requirements.txt
selenium
transformers
Pillow
numpy
opencv-python